Hero image for tags page
The styles work ideally for a smaller site such as this one. I’m able
to add the variant image sizes for the main pages with heroes, but not
the individually posts. Having variant background image URLs based on
viewports does require hardcoded values and separate classes. The
individual posts have an inline URL for bg images, but its optional and
only on the audio posts. Those are a single file, which can be
compressed and have been sitting around ~100KB which isn’t too bad.